backend/requirements.txt
fastapi
uvicorn
pydantic
motor
joblib
pandas
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
openai
uuid
python-multipart
aiosmtplib
